+++ web/src/components/Providers.jsx
@@ -1,5 +1,5 @@
 import { useEffect, useMemo, useState } from "react";
-import * as Dialog from "@radix-ui/react-dialog";
+import * as Dialog from "./ResponsiveDialog.jsx";
 import { Command } from "cmdk";
 import PageFrame from "./PageFrame.jsx";
 import { api } from "../lib/api.js";