+++ web/src/components/WorkspaceAttach.jsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import * as Dialog from "@radix-ui/react-dialog";
+import * as Dialog from "./ResponsiveDialog.jsx";
 import { api } from "../lib/api.js";
 import { IconFile, IconFolder } from "./Icons.jsx";